requirements.txt
flask
pandas
requests
beautifulsoup4
lxml
flask-sqlalchemy
flask-login
flask-socketio
flask-wtf
eventlet
openpyxl
apscheduler
python-dotenv
gunicorn
